requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12
pdfplumber==0.11.4
cryptography==43.0.0
python-dotenv==1.0.1
requests==2.32.3
fpdf2==2.7.9
segno==1.6.1
